Retrospectively stickerize to fix #5

+++ handler/image.go
@@ -5,6 +5,13 @@ import (
 	"errors"
 	"fmt"
 	"image"
+
+	// Import all possible image codecs
+	// for getImageDimensions
+	_ "image/gif"
+	_ "image/jpeg"
+	_ "image/png"
+
 	"mime"
 	"net/http"
 	"os"
@@ -43,6 +50,9 @@ func (handler *Image) SetUp(client *whatsmeow.Client, event *events.Message) {
 }
 
 func (handler *Image) Validate() error {
+	if handler == nil {
+		return errors.New("Please initialize handler")
+	}
 	event := handler.Event
 	image := event.Message.GetImageMessage()
 	if image.GetFileLength() > ImageFileSizeLimit {
@@ -59,6 +69,9 @@ func (handler *Image) Validate() error {
 }
 
 func (handler *Image) Handle() *waProto.Message {
+	if handler == nil {
+		return nil
+	}
 	// Download Image
 	event := handler.Event
 	image := event.Message.GetImageMessage()
@@ -116,6 +129,9 @@ func (handler *Image) Handle() *waProto.Message {
 }
 
 func (handler *Image) SendResponse(message *waProto.Message) {
+	if handler == nil {
+		return
+	}
 	event := handler.Event
 	completed := &waProto.Message{
 		ExtendedTextMessage: &waProto.ExtendedTextMessage{
@@ -128,11 +144,16 @@ func (handler *Image) SendResponse(message *waProto.Message) {
 }
 
 func (handler *Image) CleanUp() {
+	if handler == nil {
+		return
+	}
 	os.Remove(handler.RawPath)
 	os.Remove(handler.ConvertedPath)
 
 }
 
+// FIXME: Probably use this image dimensions to find a way
+// to maintain aspect ratio on an image before uploading as sticker
 func getImageDimensions(path string) (int, int) {
 	width, height := 0, 0
 	if reader, err := os.Open(path); err == nil {
